Fixes #10
This commit sets the MemcardA/MemcardB setting in Dolphin dynamically,
based on the current environment. Before this commit, the value was
static, and set inside of the zip file the emulator is downloaded from.
Now, Ice will figure out the location that it should be on disk, and
change Dolphin.ini accordingly