Uses registry for Steam Location. Fixes #24
Apparently, Steam stores the path to itself in the registry. This makes
sense given that Steam needs to be able to find its installation
directory for the same reason I do, and it would be super easy for the
installer to just throw the location in the registry. I digress, Steam
stores its location in the registry. Previously, I required the user to
enter the location of their Steam installation. This commit removes that
requirements and gets rid of the location option in config.txt entirely