Adds ROM directory to config.txt. Fixes #23
Adds another section to config.txt, called 'Storage'. Inside Storage,
there is an option to specify the location which Ice uses to look for
ROMs. This path can be anything, as long as it is writable (so that I
can throw the ROMs directory and subdirectories there). If the directory
doesn't exist, or isn't writable, then a default of ~ is used.

Also adds information in config.txt to explain this to users