Allows user to specify emulators in config.txt
Previously the emulators for a given system were hard coded. It is silly
to think that those emulators will work well for everyone though, so to
fix this, the user can now specify which emulator they want to use for a
system inside of config.txt. These emulators need to be supported by Ice
(aka have a subclass of Emulator dedicated to them), so it isnt just a
drag and drop process.

My goal is to continuously add new emulators and others who
want to use Ice can add them as well.