Inlines the array of supported consoles
Previously generating the list of supported consoles had two steps. The
first was to create an object, and the second was to create the array.
The array is the only important thing, and we ended up creating objects
for consoles we didnt even use.

This commit fixes the issue by instead inlining all of the console
creation inside the supported_consoles array. Then the objects for
consoles like 'Wii' (which I dont support at time of writing) will never
be created and waste our time.