Updates SteamShortcutManager. Fixes #29
See the relevant SteamShortcutManager commit:
https://github.com/scottrice/SteamShortcutManager/commit/80f63756f18e0be2bd99be1ca0ef1ab087dbbea3

The issue was caused because I use a recursive decent parser to parse
the shortcuts.vdf file. As was pointed out, this causes a problem when
there are more shortcuts than pythons recursion limit. To fix this, I
instead put the array parsing in a `while True` loop. This loop just
updates the solitary parameter with the value I was recursing with
anyway. Then, when we dont find any more matches, we know we have hit
our base case so we just return the array we have been building up.