Fixes a crash on Mac/Linux
I apparently used to have destdir as a global, but I removed it at some
point. On Mac/Linux, the 'execute' permission needs to be set on the
binary, but not on Windows. As such, I only call this function on
Mac/Linux, hence why it has gone unnoticed for so long. This commit
fixes the issue by passing the destdir variable as a parameter to the
function.