Creates Ice appdata directory if needed. Fixes #46
If the app data directory doesn't exist, then Ice can't place a log
file. This didnt use to be a problem, because emulators were downloaded
the minute Ice was run, and there is code to ensure that there is a
location to download to. When I changed the emulators to only download
when they were needed, it ended up pushing this directory creation back.
As such, the first time Ice goes to log something (which as of now is
when creating the ROMs directory) we get an error that the directory
cannot be found.

I solve this problem by creating the app data directory when the
filesystem helper is imported.