Removes deprecated functionality. Fixes #72
Removes everything related to the 'script' method of launching an
emulator. This method involved creating a script in the 'Exes' directory
for a game and then asking Steam to execute the script.

Removes emulator subclasses that corresponded to specific emulators. In
the new system of emulators.txt and consoles.txt, these methods should
not be necessary.

Removes every reference to the 'app_data_directory', which is no longer
used and therefore no longer needed.

Removes the automatic creating of ROMs directory on importing of
roms_manager. Instead, this step is explicit in ice.py.

Removes unit tests which are no longer relevant