More conservative on which shortcuts it deletes.
Previously, Ice assumed that any shortcut that pointed to the ROMs
directory was managed by Ice. That meant that if a shortcut poitned to
the ROMs directory, but Ice wouldnt add that shortcut itself, it deleted
it. This was a problem for people who manually set up some shortcuts
outside of Ice and stored the ROMs in the same ROMs directory.

We fix this by being more specific with the 'is_managed_by_ice'
function. Instead of just looking for the ROMs directory in the path,
the function now loops through all consoles and looks for their specific
ROMs directory.

One side effect of this change is that if you remove a console from Ice
(instead of removing the ROMs), all of the shortcuts will stay. I'm not
sure if this is a bug or a feature.

Fixes #109