Expands ~ in emulator location. Fixes #131
An emulators location is grabbed from the config file, but when I just
stored it directly as a string. When I went to build the command string
I made sure to do os.path.expanduser, but I forgot to do so in the
startdir. To fix this, I call os.path.expanduser as soon as I read
the location from config so that it is formatted correctly every time I
reference it later.