Adds prefix option for ROM titles in consoles.txt
Adds an optional parameter to consoles.txt, called 'prefix'. When naming
ROMs in Steam for a console, the value of the 'prefix' field will be
prepended to the name, along with a space. For example, if the prefix
was '[N64]' and the ROM name was 'Smash Bros', the final name in Steam
would be '[N64] Smash Bros'. If the value of the 'prefix' field is the
empty string, then no space is appended.

Fixes #116