Adds ability to get grid images locally
Splits grabbing of grid images into two classes, LocalProvider and
ConsoleGridProvider, along with some architecture changes to allow an
unlimited number of providers.

LocalProvider looks in a consoles 'images
directory' (which can be specified in consoles.txt) for an image with
the same name as the ROM. If one exists, it will use that

ConsoleGridProvider does what Ice has been doing before, queries
ConsoleGrid to see if an image has been uploaded for the game. If so,
will download the image and use that.

Fixes #106