Renames the prefix-related methods in rom.py
Originally when prefixes were added, rom.name() would give the ROM name
including the prefix, and rom.basename() had to be used to get it
without the prefix. I decided that the prefix is only really useful in
one situation (creating the shortcut) so it made no sense to hijack a
method which had been in use for such a long time. Instead, rom.name()
behaves as it used to, and the prefix related code has been moved to
rom.prefixed_name(). rom.to_shortcut() has been changed accordingly