Fixes reference to unused variable in error cases
When ConsoleGrid had a game, but didn't have an image set for it, Ice
would try to throw an error that included the name of the game when the
ROM object hadnt been passed to the function. To fix this, I have
changed around the function calls such that download_image now takes a
ROM and downloads an image for that ROM (as opposed to the URL it took
before).

Fixes #150