Ice can now remove old ROMs after a ROMs directory change
Getting this working invovled a lot of learning about Steam. Mainly, there are two things:

1) Steam allows you to enter an arbitrary number of tags, but only displays 1 of them.
2) The tag that Steam displays seems to be whichever comes first alphabetically (may be special for the 'favorite' tag)
3) This alphabetical search is case insensitive.

So to get a tag that I was sure would never show up, I just made it start with a character that comes after all the letters in the alphabet. In this case, I went with `~`. The ~ at the front of the flag ensures that the console name will always be shown before the `~ManagedByIce` tag.