Adds CombinedProvider to aggregate results from multiple different grid providers
The logic to determine the hierarchy of grid provider results was stuck in ROMManager, which was kind of a catch-all class for syncing functionality before. I would like to move this functionality into its own object for testability, hence CombinedProvider. CombinedProvider implements the GridImageProvider interface itself, but merely provides a wrapper of that interface over multiple different objects. It is order-dependent, so `CombinedProvider(provider1, provider2)` will give different results from `CombinedProvider(provider2, provider1)`. This has the benefit though of making consumers completely unaware of whether they are interacting with a group of providers or with a single one.