Adds a class to handle the Shortcut synchronization logic
This logic used to be housed in ROMManager, which despite being a super generic name was also not really a great place for it. ROMManager did backups, syncing, and artwork updating.

In an earlier commit I removed the artwork updating functionality from that class, and in this commit I remove the backups and syncing. Backups are now a part of IceEngine (which I think is where they should be) and Syncing is now part of this new SteamShortcutSynchronizer object. This new object is in charge of:

* Determining if a shortcut on the Users system is managed by Ice
* Determining the shortcuts that were added
* Determining the shortcuts that were removed
* Syncing a users Shortcuts with a list of ROMs

This also means that ROMManager can die. Its one of the oldest pieces of code in Ice. RIP.