Adds a script to install a set of git hooks for this project
The script just makes a symlink from the .git/hooks directory to the tools/git-hooks directory, meaning that if you use the install script that the hooks will be in version control. That is super handy for me, as I can add things like linters to the pre-commit hooks.

As of right now the hooks simply run Unit Tests before pushing. If unit tests fail it will block the push.