Makes BackedObject's immutable
I want to move to an immutable data model, cause that is what all the cool kids on the block are doing. To do so, I need to modify my BackedObject class to be immutable. Doing so also requires me to implement the save functionality somewhere else. In this case, I feel that it fits best with the BackedObjectManager.