Removes the `roms_directory` function of Console.
I want to make Console and other objects as dumb as possible. As such, one of the things I need to do is get rid of consoles ability to figure out its own roms directory. It should know if it has an override, but it shouldn't need to go to the config and figure out what the roms directory is for that console. Some other object should do that.