Removes the necessity of `initialize` from BackedObjectManager
BackedObjectManager used to have an `initialize` function that would actually read all of the values from the store and cache it. The entire reason this happened was because I wanted to initialize the ConsoleManager before the EmulatorManager but Consoles rely on data from Emulators. I dont know why I had such a huge issue with ordering those commits differently, but the last commit reordered those initailizers so now there is no reason for `initialize`.

To keep the initializer of BackedObjectManager dumb, I instead create the objects lazily when they are requested by `find`. This takes advantage of the fact that BackedObjectManager knows ahead of time every object that it SHOULD be able to return, so it can return None immediately on objects it shouldnt and cache the objects it creates otherwise.