Removes subclasses of BackedObjectManager in favor of composition
Instead of making subclasses of BackedObjectManager just to tell it how to create a BackedObject I instead pass in a BackedObjectAdapter. This adapter will contain only the BackedObject specific code, allowing me to reuse the main logic of BackedObjectManager without needing to worry about subclasses calling super in the correct places.