Adds ManagedROMArchive to check whether ROM was added by Ice
ManagedROMArchive is in charge of checking whether a ROM was managed by Ice in the last run. This will let Ice be more confident on whether to delete a random shortcut. Right now it only handles READING from an archive, assuming that the contents of the archive are a JSON string that describes a `userid32 => appid array` map. If the file is missing, or if there is no entry for a user in that file, then an empty list is assumed (aka - this is the first time Ice is being run for this user/environment). If the contents of the file are malformed it will throw a ValueError (or, more specifically, the JSON decoder will).