Fix issue with pip > 6.0
pip 6.0 added a required argument to the `parse_requirements` method. This commit adds in a value for that keyword

To test this I ran `pip install -U pip` to get a version > 6.0. I ran setup.py and confirmed that it worked. Then I downgraded back to my previous version (1.5.6) by using `pip install -U pip==1.5.6` and reran setup.py. Confirmed that it worked.

Fixes #252 and #255