Fixes an issue that was stopping us from hitting ConsoleGrid for images
This was due to an oversight on my part with the fact that the CombinedProvider didn't handle the exception case well. Since LocalProvider was throwing an exception if there was no images directory then not specifying that directory would stop us from hitting ConsoleGrid at all.

To fix this I removed the concept of ProviderError. It wasn't really providing much use (it just raised the errors to immediatly be sent to a debug log) so instead I just log explicitly in their place. This makes control flow much simpler, and so now CombinedProvider should handle these cases correctly.