Remove level from INFO logs on Console, improve file logging
Adds a bunch of new fields to the file logging to help out with debugging, along with removing the "INFO" from all of the regular logs to console. Now we only show the level if it is a warning or above. This means I can use the logger to log ACTUAL info, like if I want to print out a list of things that the user needs to fix.

Fixes #265