Checks non-user specific environment once instead of once per user
This keeps clowny things from happening like "Running for User X: cant run because Steam is open. Running for User Y: cant run because Steam is open. Running for User Z: cant run because Steam is open". This should just show that message once.

It unfortunately complicates the EnvChecker code in IceEngine considerably, so I will need to refactor that.