Stop auto adding `ICE_FLAG_TAG` to every shortcut. Fixes #284.
Ice tries to //sync// ROMs, but figuring out which ROMs we should delete has always been tough. At one point I tried a heuristic ('does it look like a shortcut Ice would make?'), but that was extremely error prone (as you might imagine) so I went with adding a special tag that Steam didn't show to every game. That was a terrible idea, since it relied on an implementation detail of Steam and ended up being user facing. Now that we have a history-based system I can just straight remove this tag from all of our games. FINALLY.

ROMs that have the tag will still be flagged as Managed if it exists though.