Differentiate between 'No history' and 'Emtpy history' in the archive
Now that we rely on the history to tell us if a shortcut is 'managed' or not we are in an unfortunate situation where people that have used Ice previously will get duplicates of all of their games. This is because there is no record kept from when Ice used the heuristic based method. To allow for migration from one system to the other, we need to differentiate between the two cases (no history vs 'we have a history but the user didn't do anything last time')