Add an extremely basic integration test
Since the inputs to Ice are the state of the users filesystem, it can be REALLY hard to run a full manual test suite. I've been worried about adding new features recently because I was worried that I was breaking things, and unless I did some really intense work I couldn't tell if something broke or not in one of the variety of different edge cases.

Well now I have the infra necessary for integration tests, so I can be sure that Ice still actually runs without syntax erroring in the most common case.

This is mainly done through the CommandLineRunner, and I'm modeling the API after what Buck does (you give it a set of CLI args).