Convert emulator into a value object
Code should be thought about in terms of data and functionality, not a mix. Plus I didnt like the fact that the console/emulator objects werent just dumb data representing the contents of consoles.txt and emulators.txt.

This is a holdover from the days when I had a master list somewhere of all of the Consoles/Emulators that Ice supported. Thats no longer a thing, so the days of having these complex objects representing this data is way past gone.

The killer feature of this change is that now all of the sanity checking for the emulator (whether the executable that we expect exists) happens in the EmulatorBackedObjectAdapter, which we can pass all sorts of objects to. The emulator object is too low level, and you shouldn't need an instance of a filesystem just to create an emulator object