Change semantics of the backup option.
I need an easy way to disable backups for certain tests (but not others), but I also want to give users a sane backups directory by default. I can't have both of those options be "Leave blank" so the solution I came up with was "If you dont have the key set, we wont back up. If you have the option but leave it blank, we will choose a sane backups directory. If you explicitly chose a directory then we use that".