Replace PFA logger with one that can be just imported+used.
A logger object is great when you have objects which can hold onto an instance, but when you're writing pure functions its annoying to pass in a logger object.

A common python idiom (or so it seems) is to have a module which instantiates a logger on import, which can then be directly used. Ice itself used this approach for a while, until I removed it.