Consolidate model validation logic in model adapters
For a while I've been debating on what the best location for the `is_enabled` logic would be. Previously I just had them as free functions, but the fact that there was both a `console_is_enabled` and `emulator_is_enabled` function signaled to me that something was wrong.

Well I want to move more logic out of the environment checker, so I was wondering where I could consolidate the "create directory for console if it needs it" and the "if the console doesnt have an emulator we shouldnt include it". I decided to unify it in the concept of a "verifier". My thought process was that if we would skip a console because it didnt have any emulators, it seems reasonable that if we couldn't ensure a decent directory structure for a console then we should skip it too.

I prefer this to just checking to see if a ROMs directory exists before starting the ROM finder.