Remove `console_manager` and `emulator_manager` from Configuration
This is needed for my continued work to turn Configuration into a value object. Theres no reason that our bag of strings/bools/paths needs to know about the consoles/emulators, and it was really odd that that object did the wiring.

Instead, that logic has been moved into the new `settings` namespace, which will contain a lot of the top level wiring between user-specified values and full python objects.