Dont print to sdout in unit tests.
Since the logs module was imported early it was using the wrong `sys.stderr`, so you would still get logspew in stderr during the test, which throws off the test reporting.

This commit uses the inspect module to figure out if we're running in a unit test and doesnt add the stream/file handler to the logger when that happens, meaning only the capture handler gets used.