Correctly parse ROMs with periods in their names (Fixes #364)
I originally set up the regex so that it would stop matching at `.` characters, but in hindsight that was a dumb idea. The goal was to not matching file extensions, but python already has a perfectly good `splitext` function that removes the extension. All I had to do to get the functionality I wanted was to remove the period as a stop character and only run the regex over the filename, not the filename + extension.